Move also TestResult in report package
