+++ src/main/java/io/vertx/ext/unit/TestSuite.java
@@ -3,12 +3,9 @@ package io.vertx.ext.unit;
 import io.vertx.codegen.annotations.Fluent;
 import io.vertx.codegen.annotations.GenIgnore;
 import io.vertx.codegen.annotations.VertxGen;
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.impl.TestSuiteImpl;
-import io.vertx.ext.unit.report.ReportOptions;
 
 import java.util.concurrent.TimeUnit;
