+++ src/main/java/io/vertx/ext/unit/impl/EventBusAdapterImpl.java
@@ -7,7 +7,7 @@ import io.vertx.core.streams.ReadStream;
 import io.vertx.ext.unit.EventBusAdapter;
 import io.vertx.ext.unit.Failure;
 import io.vertx.ext.unit.report.TestCaseReport;
-import io.vertx.ext.unit.TestResult;
+import io.vertx.ext.unit.report.TestResult;
 import io.vertx.ext.unit.report.TestSuiteReport;
 
 /**
