+++ src/main/java/io/vertx/ext/unit/impl/TestCaseReportImpl.java
@@ -2,7 +2,7 @@ package io.vertx.ext.unit.impl;
 
 import io.vertx.core.Handler;
 import io.vertx.ext.unit.Test;
-import io.vertx.ext.unit.TestResult;
+import io.vertx.ext.unit.report.TestResult;
 import io.vertx.ext.unit.report.TestCaseReport;
 
 /**
