+++ src/main/java/io/vertx/ext/unit/impl/TestSuiteImpl.java
@@ -6,7 +6,7 @@ import io.vertx.ext.unit.TestCompletion;
 import io.vertx.ext.unit.TestOptions;
 import io.vertx.ext.unit.TestSuite;
 import io.vertx.ext.unit.Test;
-import io.vertx.ext.unit.TestResult;
+import io.vertx.ext.unit.report.TestResult;
 import io.vertx.ext.unit.TestSuiteRunner;
 import io.vertx.ext.unit.report.Reporter;
