+++ src/main/java/io/vertx/ext/unit/report/Reporter.java
@@ -1,7 +1,6 @@
 package io.vertx.ext.unit.report;
 
 import io.vertx.core.Vertx;
-import io.vertx.ext.unit.TestResult;
 import io.vertx.ext.unit.report.impl.DefaultReporterFactory;
 
 /**
