+++ src/main/java/io/vertx/ext/unit/report/TestResult.java
@@ -1,7 +1,8 @@
-package io.vertx.ext.unit;
+package io.vertx.ext.unit.report;
 
 import io.vertx.codegen.annotations.CacheReturn;
 import io.vertx.codegen.annotations.VertxGen;
+import io.vertx.ext.unit.Failure;
 
 /**
  * The result of a test.
